made a reservation form component, passed to newreservation
